backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
stripe==8.0.0
sentry-sdk==1.40.0
psutil==5.9.0
psycopg2-binary==2.9.9
Pillow==11.0.0
python-decouple==3.8
PyJWT==2.12.1
djangorestframework-simplejwt==5.5.1
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
django-celery-email==3.0.0
django-otp==1.5.0
django-redis==6.0.0
